pyproject.toml
[tool.poetry]
name = "product-management-backend"
version = "0.0.0"
description = ""
authors = ["Edgar Valdez <edgarvaldez99@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.2.2"
crypto = "^1.4.1"
psycopg2 = "^2.9.7"
psycopg2-binary = "^2.9.7"
Pillow = "^10.0.0"
django-cors-headers = "^4.2.0"
drf-extra-fields = "^3.7.0"

[tool.poetry.dev-dependencies]
types-requests = "^2.31.0"
mypy = "^1.5.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
isort = "^5.12.0"
black = "^23.7.0"
pre-commit = "^3.3.3"
nox = "^2023.4.22"
coverage = "^7.3.0"

[tool.mypy]
exclude = "manage.py"
ignore_missing_imports = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
